Implement GitHub issue therealsiege/Penpal#335: Extract animation state
